+++ components/tcf/services/src/application/Service.js
@@ -3,15 +3,17 @@ class Service {
     getVendorListUseCase,
     loadUserConsentUseCase,
     saveUserConsentUseCase,
+    updateUserConsentUseCase,
     uiVisibleUseCase
   }) {
     this._getVendorListUseCase = getVendorListUseCase
     this._loadUserConsentUseCase = loadUserConsentUseCase
     this._saveUserConsentUseCase = saveUserConsentUseCase
+    this._updateUserConsentUseCase = updateUserConsentUseCase
     this._uiVisibleUseCase = uiVisibleUseCase
   }
 
-  getVendorList({language}) {
+  getVendorList({language = 'es'} = {}) {
     return this._getVendorListUseCase.execute({language})
   }
 
@@ -19,11 +21,25 @@ class Service {
     return this._loadUserConsentUseCase.execute()
   }
 
-  saveUserConsent({purpose, vendor, specialFeatures}) {
-    return this._saveUserConsentUseCase.execute({
+  saveUserConsent() {
+    return this._saveUserConsentUseCase.execute()
+  }
+
+  updateUserConsent({
+    purpose,
+    vendor,
+    specialFeatures,
+    allPurposes,
+    allVendors,
+    allSpecialFeatures
+  }) {
+    return this._updateUserConsentUseCase.execute({
       purpose,
       vendor,
-      specialFeatures
+      specialFeatures,
+      allPurposes,
+      allVendors,
+      allSpecialFeatures
     })
   }