+++ components/tcf/services/src/infrastructure/Tcf/factory.js
@@ -1,8 +1,0 @@
-import {TcfRepository} from './TcfRepository'
-import TcfApiInitializer from '@adv-ui/boros-tcf'
-
-const tcfApi = TcfApiInitializer.init()
-
-export function tcfRepositoryFactory() {
-  return new TcfRepository({tcfApi, window})
-}